refactor: Dashboard JS split: extract diagnostics.js (per-domain split,… [ORB-00175]
Dashboard JS split: extract diagnostics.js (per-domain split, round 4)
